Got rid of one enemy, increased speed, coded buttons
I got rid of the fourth enemy because it would be too hard to play the
game with the increased speed.  I had to increase the speed by
decreasing the delay so the meggy would read the buttons better.  I
also programmed the basket to move when the left and right buttons are
pressed.